add change modal styles
