+++ src/partials/modal-backdrop.html
@@ -26,15 +26,33 @@
   <ul class="franchise__form">
     <li class="modal-form__item">
       <label class="modal-form__label" for="name">Name</label>
-      <input class="modal-form__input" type="text" name="name" id="name" />
+      <input
+        class="modal-form__input"
+        type="text"
+        name="name"
+        id="name"
+        placeholder=" "
+      />
     </li>
     <li class="modal-form__item">
       <label class="modal-form__label" for="tel">Phone</label>
-      <input class="modal-form__input" type="tel" name="tel" id="tel" />
+      <input
+        class="modal-form__input"
+        type="tel"
+        name="tel"
+        id="tel"
+        placeholder=" "
+      />
     </li>
     <li class="modal-form__item">
       <label class="modal-form__label" for="mail">Email</label>
-      <input class="modal-form__input" type="email" name="mail" id="mail" />
+      <input
+        class="modal-form__input"
+        type="email"
+        name="mail"
+        id="mail"
+        placeholder=" "
+      />
     </li>
     <li class="modal-form__item">
       <label class="modal-form__label" for="comment">Comment</label>
